# Use structured wheel tags everywhere