--- a/crates/uv-build-backend/src/wheel.rs
+++ b/crates/uv-build-backend/src/wheel.rs
@@ -1,20 +1,23 @@
-use crate::metadata::{BuildBackendSettings, DEFAULT_EXCLUDES};
-use crate::{DirectoryWriter, Error, FileList, ListWriter, PyProjectToml};
+use std::io::{BufReader, Read, Write};
+use std::path::{Path, PathBuf};
+use std::{io, mem};
+
 use fs_err::File;
 use globset::{GlobSet, GlobSetBuilder};
 use itertools::Itertools;
 use sha2::{Digest, Sha256};
-use std::io::{BufReader, Read, Write};
-use std::path::{Path, PathBuf};
-use std::{io, mem};
 use tracing::{debug, trace};
-use uv_distribution_filename::WheelFilename;
+use walkdir::WalkDir;
+use zip::{CompressionMethod, ZipWriter};
+
+use uv_distribution_filename::{TagSet, WheelFilename};
 use uv_fs::Simplified;
 use uv_globfilter::{parse_portable_glob, GlobDirFilter};
 use uv_platform_tags::{AbiTag, LanguageTag, PlatformTag};
 use uv_warnings::warn_user_once;
-use walkdir::WalkDir;
-use zip::{CompressionMethod, ZipWriter};
+
+use crate::metadata::{BuildBackendSettings, DEFAULT_EXCLUDES};
+use crate::{DirectoryWriter, Error, FileList, ListWriter, PyProjectToml};
 
 /// Build a wheel from the source tree and place it in the output directory.
 pub fn build_wheel(
@@ -34,12 +37,12 @@ pub fn build_wheel(
         name: pyproject_toml.name().clone(),
         version: pyproject_toml.version().clone(),
         build_tag: None,
-        python_tag: vec![LanguageTag::Python {
+        python_tag: TagSet::from_slice(&[LanguageTag::Python {
             major: 3,
             minor: None,
-        }],
-        abi_tag: vec![AbiTag::None],
-        platform_tag: vec![PlatformTag::Any],
+        }]),
+        abi_tag: TagSet::from_buf([AbiTag::None]),
+        platform_tag: TagSet::from_buf([PlatformTag::Any]),
     };
 
     let wheel_path = wheel_dir.join(filename.to_string());
@@ -72,12 +75,12 @@ pub fn list_wheel(
         name: pyproject_toml.name().clone(),
         version: pyproject_toml.version().clone(),
         build_tag: None,
-        python_tag: vec![LanguageTag::Python {
+        python_tag: TagSet::from_slice(&[LanguageTag::Python {
             major: 3,
             minor: None,
-        }],
-        abi_tag: vec![AbiTag::None],
-        platform_tag: vec![PlatformTag::Any],
+        }]),
+        abi_tag: TagSet::from_buf([AbiTag::None]),
+        platform_tag: TagSet::from_buf([PlatformTag::Any]),
     };
 
     let mut files = FileList::new();
@@ -254,12 +257,12 @@ pub fn build_editable(
         name: pyproject_toml.name().clone(),
         version: pyproject_toml.version().clone(),
         build_tag: None,
-        python_tag: vec![LanguageTag::Python {
+        python_tag: TagSet::from_slice(&[LanguageTag::Python {
             major: 3,
             minor: None,
-        }],
-        abi_tag: vec![AbiTag::None],
-        platform_tag: vec![PlatformTag::Any],
+        }]),
+        abi_tag: TagSet::from_buf([AbiTag::None]),
+        platform_tag: TagSet::from_buf([PlatformTag::Any]),
     };
 
     let wheel_path = wheel_dir.join(filename.to_string());
@@ -309,12 +312,12 @@ pub fn metadata(
         name: pyproject_toml.name().clone(),
         version: pyproject_toml.version().clone(),
         build_tag: None,
-        python_tag: vec![LanguageTag::Python {
+        python_tag: TagSet::from_slice(&[LanguageTag::Python {
             major: 3,
             minor: None,
-        }],
-        abi_tag: vec![AbiTag::None],
-        platform_tag: vec![PlatformTag::Any],
+        }]),
+        abi_tag: TagSet::from_buf([AbiTag::None]),
+        platform_tag: TagSet::from_buf([PlatformTag::Any]),
     };
 
     debug!(
@@ -766,7 +769,7 @@ mod test {
             name: PackageName::from_str("foo").unwrap(),
             version: Version::from_str("1.2.3").unwrap(),
             build_tag: None,
-            python_tag: vec![
+            python_tag: TagSet::from_slice(&[
                 LanguageTag::Python {
                     major: 2,
                     minor: None,
@@ -775,9 +778,9 @@ mod test {
                     major: 3,
                     minor: None,
                 },
-            ],
-            abi_tag: vec![AbiTag::None],
-            platform_tag: vec![PlatformTag::Any],
+            ]),
+            abi_tag: TagSet::from_buf([AbiTag::None]),
+            platform_tag: TagSet::from_buf([PlatformTag::Any]),
         };
 
         assert_snapshot!(wheel_info(&filename, "1.0.0+test"), @r"