… Cargo.lock
… crates/uv-build-backend/Cargo.toml
… crates/uv-distribution-filename/Cargo.toml
… crates/uv-distribution-filename/src/snapshots/uv_distribution_filename__wheel__tests__ok_build_tag.snap
… crates/uv-distribution-filename/src/snapshots/uv_distribution_filename__wheel__tests__ok_multiple_tags.snap
… crates/uv-distribution-filename/src/snapshots/uv_distribution_filename__wheel__tests__ok_single_tags.snap
… crates/uv-platform-tags/Cargo.toml
… crates/uv-platform-tags/src/language_tag.rs
… crates/uv-platform-tags/src/lib.rs
… crates/uv-platform-tags/src/platform.rs
… crates/uv-platform-tags/src/platform_tag.rs
… crates/uv-platform-tags/src/tags.rs
… crates/uv-publish/src/lib.rs
… crates/uv-python/src/platform.rs
… crates/uv-resolver/src/lock/mod.rs
… crates/uv-resolver/src/lock/snapshots/uv_resolver__lock__tests__hash_optional_missing.snap
… crates/uv-resolver/src/lock/snapshots/uv_resolver__lock__tests__hash_optional_present.snap
… crates/uv-resolver/src/lock/snapshots/uv_resolver__lock__tests__hash_required_present.snap
… crates/uv-resolver/src/pubgrub/report.rs
… crates/uv-resolver/src/requires_python.rs
… crates/uv/tests/it/cache_clean.rs
… crates/uv/tests/it/pip_compile.rs
… crates/uv/tests/it/pip_install_scenarios.rs
… crates/uv/tests/it/pip_sync.rs
… crates/uv/tests/it/snapshots/it__ecosystem__warehouse-lock-file.snap