@charliermarsh 2025-01-12
Hmm... I wonder if the `wheelname_parsing` regressions are real. The thing is, this doesn't allocate anymore, whereas the previous implementation allocated multiple times (strings and vectors). I'll have to benchmark it on my own machine.