@charliermarsh 2025-01-13
Yeah. I need to run them locally. We are doing a bit more work, but we’re also not allocating anymore…