@charliermarsh 2025-01-13
I don't think the CodSpeed regressions are "real".

On main:

```
wheelname_parsing/flyte-short-incompatible
                        time:   [159.47 ns 160.13 ns 160.82 ns]
                        thrpt:  [201.63 MiB/s 202.49 MiB/s 203.33 MiB/s]
                 change:
                        time:   [-3.0123% -1.8530% -0.6083%] (p = 0.00 < 0.05)
                        thrpt:  [+0.6120% +1.8880% +3.1059%]
                        Change within noise threshold.
wheelname_parsing/flyte-short-compatible
                        time:   [159.66 ns 161.03 ns 162.68 ns]
                        thrpt:  [175.87 MiB/s 177.67 MiB/s 179.20 MiB/s]
                 change:
                        time:   [-2.2943% -1.2318% -0.1728%] (p = 0.02 < 0.05)
                        thrpt:  [+0.1731% +1.2472% +2.3481%]
                        Change within noise threshold.
wheelname_parsing/flyte-long-incompatible
                        time:   [357.67 ns 358.70 ns 359.72 ns]
                        thrpt:  [342.00 MiB/s 342.97 MiB/s 343.96 MiB/s]
                 change:
                        time:   [-2.3844% -2.0267% -1.6664%] (p = 0.00 < 0.05)
                        thrpt:  [+1.6946% +2.0686% +2.4427%]
```

On this branch:

```
wheelname_parsing/flyte-short-incompatible
                        time:   [119.45 ns 119.72 ns 120.01 ns]
                        thrpt:  [270.18 MiB/s 270.84 MiB/s 271.46 MiB/s]
                 change:
                        time:   [-1.2435% -0.9227% -0.5367%] (p = 0.00 < 0.05)
                        thrpt:  [+0.5396% +0.9313% +1.2592%]
                        Change within noise threshold.
wheelname_parsing/flyte-short-compatible
                        time:   [119.13 ns 119.40 ns 119.70 ns]
                        thrpt:  [239.01 MiB/s 239.61 MiB/s 240.17 MiB/s]
                 change:
                        time:   [-1.9422% -1.7395% -1.5472%] (p = 0.00 < 0.05)
                        thrpt:  [+1.5715% +1.7703% +1.9807%]
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) low severe
  4 (4.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe
wheelname_parsing/flyte-long-incompatible
                        time:   [389.86 ns 390.80 ns 392.03 ns]
                        thrpt:  [313.82 MiB/s 314.80 MiB/s 315.56 MiB/s]
                 change:
                        time:   [-0.4042% +0.1993% +0.8119%] (p = 0.52 > 0.05)
                        thrpt:  [-0.8054% -0.1989% +0.4059%]
                        No change in performance detected.
wheelname_parsing/flyte-long-compatible
                        time:   [282.53 ns 283.03 ns 283.65 ns]
                        thrpt:  [383.28 MiB/s 384.12 MiB/s 384.81 MiB/s]
                 change:
                        time:   [+0.9606% +1.1332% +1.3430%] (p = 0.00 < 0.05)
                        thrpt:  [-1.3252% -1.1205% -0.9514%]
```

After the #10546 PR:

```
wheelname_parsing/flyte-short-incompatible
                        time:   [111.87 ns 112.46 ns 112.95 ns]
                        thrpt:  [287.07 MiB/s 288.31 MiB/s 289.85 MiB/s]
                 change:
                        time:   [-0.1758% +0.5279% +1.2135%] (p = 0.12 > 0.05)
                        thrpt:  [-1.1989% -0.5251% +0.1761%]
                        No change in performance detected.
Found 23 outliers among 100 measurements (23.00%)
  17 (17.00%) low severe
  2 (2.00%) high mild
  4 (4.00%) high severe
wheelname_parsing/flyte-short-compatible
                        time:   [110.01 ns 110.39 ns 110.77 ns]
                        thrpt:  [258.28 MiB/s 259.18 MiB/s 260.07 MiB/s]
                 change:
                        time:   [-0.8660% -0.1473% +0.5616%] (p = 0.69 > 0.05)
                        thrpt:  [-0.5585% +0.1475% +0.8736%]
                        No change in performance detected.
Found 24 outliers among 100 measurements (24.00%)
  16 (16.00%) low severe
  2 (2.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
wheelname_parsing/flyte-long-incompatible
                        time:   [358.74 ns 359.57 ns 360.53 ns]
                        thrpt:  [341.23 MiB/s 342.14 MiB/s 342.94 MiB/s]
                 change:
                        time:   [-0.9707% -0.5934% -0.2323%] (p = 0.00 < 0.05)
                        thrpt:  [+0.2328% +0.5969% +0.9802%]
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
wheelname_parsing/flyte-long-compatible
                        time:   [258.74 ns 258.92 ns 259.10 ns]
                        thrpt:  [419.59 MiB/s 419.90 MiB/s 420.18 MiB/s]
                 change:
                        time:   [-0.5224% -0.3401% -0.1560%] (p = 0.00 < 0.05)
                        thrpt:  [+0.1563% +0.3413% +0.5252%]
                        Change within noise threshold.
```

So the short tag case is about 30% _faster_, and the long tag case is either unchanged or 15% faster.