@charliermarsh 2025-01-13
(My guess is that we're doing more CPU work than before (due to the parsing) but no longer allocating.)