@konstin 2025-01-13
Checking with hyperfine, there is a speedup
```
$ hyperfine --warmup 2 --runs 100 --export-json tags.json "./uv-main pip compile scripts/requirements/airflow.in" "./uv-tags pip compile scripts/requirements/airflow.in"
Benchmark 1: ./uv-main pip compile scripts/requirements/airflow.in
  Time (mean ± σ):     119.6 ms ±   5.1 ms    [User: 134.8 ms, System: 111.6 ms]
  Range (min … max):   111.2 ms … 141.2 ms    100 runs
 
Benchmark 2: ./uv-tags pip compile scripts/requirements/airflow.in
  Time (mean ± σ):     114.3 ms ±   3.9 ms    [User: 129.6 ms, System: 108.6 ms]
  Range (min … max):   106.7 ms … 127.9 ms    100 runs
 
Summary
  ./uv-tags pip compile scripts/requirements/airflow.in ran
    1.05 ± 0.06 times faster than ./uv-main pip compile scripts/requirements/airflow.in
```
and the speedup is significant (welch t-test, t = 8.12, p = 6.55e-14).